python-backend/pyproject.toml
[project]
name = "pmo-digital-platform"
version = "1.0.0"
description = "PMO Digital Platform - Python FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "python-jose[cryptography]>=3.3.0",
    "gmssl>=3.2.0",
    "redis>=5.0.0",
    "celery[redis]>=5.3.0",
    "minio>=7.2.0",
    "structlog>=24.1.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "openpyxl>=3.1.0",
    "psycopg2-binary>=2.9.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "fakeredis>=2.21.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "testcontainers[postgres]>=4.0.0",
    "playwright>=1.42.0",
    "aiosqlite>=0.20.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py312"
